# docs/swagger.yaml
definitions:
...
  dokumen.AktaNotarisModel:
    properties:
      file:
        type: string
      id:
        type: string
      id_bumd:
        type: string
      keterangan:
        type: string
      nomor:
        type: string
      notaris:
        type: string
      tanggal:
        type: string
    type: object
...
  utils.DataValidationError:
    properties:
      field:
        example: email
        type: string
      message:
        example: Invalid email address
        type: string
    type: object
...
  utils.RequestError:
    properties:
      code:
        example: 422
        type: integer
      fields:
        items:
          $ref: '#/definitions/utils.DataValidationError'
        type: array
      message:
        example: Invalid email address
        type: string
    type: object
info:
  contact:
    email: lifelinejar@mail.com
    name: API Support
  description: BUMD Service Rest API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BUMD Service
  version: "1.0"
paths:
...
  /strict/bumd/{id_bumd}/akta_notaris:
    get:
      description: get data akta notaris.
      operationId: akta_notaris-index
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Halaman yang ditampilkan
        in: query
        name: page
        type: integer
      - description: Jumlah data per halaman, maksimal 5 data per halaman
        in: query
        name: limit
        type: integer
      - description: Pencarian
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dokumen.AktaNotarisModel'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: get data akta notaris
      tags:
      - Akta Notaris
    post:
      consumes:
      - multipart/form-data
      description: create data akta notaris.
      operationId: akta_notaris-create
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Nomor Akta notaris
        in: formData
        name: nomor
        required: true
        type: string
      - description: Notaris
        in: formData
        name: notaris
        required: true
        type: string
      - description: Tanggal Akta notaris
        in: formData
        name: tanggal
        required: true
        type: string
      - description: Keterangan
        in: formData
        name: keterangan
        required: true
        type: string
      - description: File
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: create data akta notaris
      tags:
      - Akta Notaris
  /strict/bumd/{id_bumd}/akta_notaris/{id}:
    delete:
      consumes:
      - application/json
      description: delete data akta notaris.
      operationId: akta_notaris-delete
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id untuk delete data akta notaris
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: delete data akta notaris
      tags:
      - Akta Notaris
    get:
      description: get data akta notaris by id.
      operationId: akta_notaris-view
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id untuk get data akta notaris
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dokumen.AktaNotarisModel'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: get data akta notaris by id
      tags:
      - Akta Notaris
    put:
      consumes:
      - multipart/form-data
      description: update data akta notaris.
      operationId: akta_notaris-update
      parameters:
      - description: Id BUMD
        format: uuid
        in: path
        name: id_bumd
        required: true
        type: string
      - description: Id untuk update data akta notaris
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Nomor Akta notaris
        in: formData
        name: nomor
        required: true
        type: string
      - description: Tanggal Akta notaris
        in: formData
        name: tanggal
        required: true
        type: string
      - description: Keterangan
        in: formData
        name: keterangan
        required: true
        type: string
      - description: File
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: boolean
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/utils.RequestError'
        "404":
          description: Data not found
          schema:
            $ref: '#/definitions/utils.RequestError'
        "422":
          description: Data validation failed
          schema:
            items:
              $ref: '#/definitions/utils.RequestError'
            type: array
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.RequestError'
      security:
      - ApiKeyAuth: []
      summary: update data akta notaris
      tags:
      - Akta Notaris
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"